# Speed up tests: 90.8s to 1.5s