--- a/cl/favorites/tests.py
+++ b/cl/favorites/tests.py
@@ -62,6 +62,7 @@ class NoteTest(SimpleUserDataMixin, TestCase, AudioTestCase):
     ]
 
     def setUp(self) -> None:
+        super().setUp()
         # Set up some handy variables
         self.note_cluster_params = {
             "cluster_id": 1,
@@ -114,12 +115,12 @@ class UserNotesTest(BaseSeleniumTest):
     ]
 
     def setUp(self) -> None:
+        super().setUp()
         get_homepage_stats.invalidate()
         self.f = NoteFactory.create(
             user__username="pandora",
             user__password=make_password("password"),
         )
-        super().setUp()
 
     @timeout_decorator.timeout(SELENIUM_TIMEOUT)
     def test_anonymous_user_is_prompted_when_favoriting_an_opinion(
@@ -423,6 +424,7 @@ class APITests(APITestCase):
 
     @classmethod
     def setUpTestData(cls) -> None:
+        super().setUpTestData()
         cls.pandora = UserProfileWithParentsFactory.create(
             user__username="pandora",
             user__password=make_password("password"),
@@ -434,14 +436,16 @@ class APITests(APITestCase):
         )
 
     def setUp(self) -> None:
+        super().setUp()
         self.tag_path = reverse("UserTag-list", kwargs={"version": "v3"})
         self.docket_path = reverse("DocketTag-list", kwargs={"version": "v3"})
         self.client = make_client(self.pandora.user.pk)
         self.client2 = make_client(self.unconfirmed.user.pk)
 
-    def tearDown(cls):
+    def tearDown(self):
         UserTag.objects.all().delete()
         DocketTag.objects.all().delete()
+        super().tearDown()
 
     async def make_a_good_tag(self, client, tag_name="taggy-tag"):
         data = {
@@ -659,6 +663,13 @@ class APITests(APITestCase):
 
 
 class RECAPPrayAndPay(SimpleUserDataMixin, PrayAndPayTestCase):
+    @classmethod
+    def setUpClass(cls):
+        super().setUpClass()
+        p = patch("cl.favorites.signals.check_prayer_pacer")
+        p.start()
+        cls.addClassCleanup(p.stop)
+
     @override_settings(ALLOWED_PRAYER_COUNT=2)
     async def test_prayer_eligible(self) -> None:
         """Does the prayer_eligible method work properly?"""
@@ -1700,7 +1711,15 @@ class PrayAndPayCheckAvailabilityTaskTests(PrayAndPayTestCase):
 class PrayerAPITests(PrayAndPayTestCase):
     """Check that Prayer API operations work as expected."""
 
+    @classmethod
+    def setUpClass(cls):
+        super().setUpClass()
+        p = patch("cl.favorites.signals.check_prayer_pacer")
+        p.start()
+        cls.addClassCleanup(p.stop)
+
     def setUp(self) -> None:
+        super().setUp()
         self.prayer_path = reverse("prayer-list", kwargs={"version": "v4"})
         self.client = make_client(self.user.pk)
         self.client_2 = make_client(self.user_2.pk)