I hate to use emoji like this, but I think you should prioritize this one, too.  It turns out the `favorites` tests were trying to tattle to PACER every time we ran a test.  After ruthlessly mocking the little snitch, it has shut up and the tests run much faster.
I've also included missing `super()` calls and whatnot, because I can't help myself.

test:
`docker exec cl-django ./manage.py test cl.favorites.tests --verbosity=0 --exclude-tag selenium --parallel=3 --timing --durations 10 --keepdb`

main:
```
Slowest test durations
----------------------------------------------------------------------
13.754s    test_get_top_prayers_by_views (cl.favorites.tests.RECAPPrayAndPay.test_get_top_prayers_by_views)
10.615s    test_get_top_prayers_by_all (cl.favorites.tests.RECAPPrayAndPay.test_get_top_prayers_by_all)
10.591s    test_get_top_prayers_by_number_and_views (cl.favorites.tests.RECAPPrayAndPay.test_get_top_prayers_by_number_and_views)
8.729s     test_get_lifetime_prayer_stats (cl.favorites.tests.RECAPPrayAndPay.test_get_lifetime_prayer_stats)
7.831s     test_get_top_prayers_by_number (cl.favorites.tests.RECAPPrayAndPay.test_get_top_prayers_by_number)
6.666s     test_prayer_eligible (cl.favorites.tests.RECAPPrayAndPay.test_prayer_eligible)
6.632s     test_get_user_prayer_history (cl.favorites.tests.RECAPPrayAndPay.test_get_user_prayer_history)
5.271s     test_total_prayer_cost_is_rounded (cl.favorites.tests.RECAPPrayAndPay.test_total_prayer_cost_is_rounded)
5.237s     test_get_user_prayers (cl.favorites.tests.RECAPPrayAndPay.test_get_user_prayers)
3.983s     test_prayers_integration (cl.favorites.tests.RECAPPrayAndPay.test_prayers_integration)

----------------------------------------------------------------------
Ran 46 tests in 90.777s

OK
```

gotta-go-fast:
```
Slowest test durations
----------------------------------------------------------------------
0.202s     test_get_user_prayers (cl.favorites.tests.RECAPPrayAndPay.test_get_user_prayers)
0.180s     test_revert_tracked_model (cl.favorites.tests.FavoritesTest.test_revert_tracked_model)
0.141s     test_can_filter_tag_associations_using_docket_id (cl.favorites.tests.APITests.test_can_filter_tag_associations_using_docket_id)
0.091s     test_create_note (cl.favorites.tests.NoteTest.test_create_note)
0.086s     test_user_gets_notification_when_document_is_unavailable (cl.favorites.tests.PrayAndPayCheckAvailabilityTaskTests.test_user_gets_notification_when_document_is_unavailable)
0.081s     test_can_filter_tag_associations_using_user_id (cl.favorites.tests.APITests.test_can_filter_tag_associations_using_user_id)
0.056s     test_get_lifetime_prayer_stats (cl.favorites.tests.RECAPPrayAndPay.test_get_lifetime_prayer_stats)
0.050s     test_list_of_granted_prayers_is_always_private (cl.favorites.tests.RECAPPrayAndPay.test_list_of_granted_prayers_is_always_private)
0.048s     test_can_we_load_the_top_prayers_page (cl.favorites.tests.RECAPPrayAndPay.test_can_we_load_the_top_prayers_page)
0.047s     test_delete_prayer (cl.favorites.tests.PrayerAPITests.test_delete_prayer)

----------------------------------------------------------------------
Ran 46 tests in 1.480s

OK
```